+++ prog/dftb+/lib_dftbplus/mainio.F90
@@ -1782,7 +1782,7 @@ contains
     !> list of region names
     type(TListCharLc), intent(inout) :: fileNames
 
-    !> eigenvalues
+    !> Eigenvalues
     real(dp), intent(in) :: eigvals(:,:,:)
 
     !> Neighbour list.
@@ -2034,9 +2034,9 @@ contains
 
 
   !> Writes out machine readable data
-  subroutine writeResultsTag(fileName, energy, derivs, chrgForces, electronicSolver, tStress,&
-      & totalStress, pDynMatrix, tPeriodic, cellVol, tMulliken, qOutput, q0, taggedWriter,&
-      & tDefinedFreeE, cm5Cont)
+  subroutine writeResultsTag(fileName, energy, derivs, chrgForces, nEl, Ef, eigen, filling,&
+      & electronicSolver, tStress, totalStress, pDynMatrix, tPeriodic, cellVol, tMulliken,&
+      & qOutput, q0, taggedWriter, tDefinedFreeE, cm5Cont)
 
     !> Name of output file
     character(*), intent(in) :: fileName
@@ -2049,6 +2049,18 @@ contains
 
     !> Forces on external charges
     real(dp), allocatable, intent(in) :: chrgForces(:,:)
+
+    !> Number of electrons
+    real(dp), intent(in) :: nEl(:)
+
+    !> Fermi level(s)
+    real(dp), intent(inout) :: Ef(:)
+
+    !> Eigenvalues/single particle states (level, kpoint, spin)
+    real(dp), intent(in) :: eigen(:,:,:)
+
+    !> Filling of the eigenstates
+    real(dp), intent(in) :: filling(:,:,:)
 
     !> Electronic solver information
     type(TElectronicSolver), intent(in) :: electronicSolver
@@ -2095,11 +2107,15 @@ contains
     open(newunit=fd, file=fileName, action="write", status="replace")
 
     call taggedWriter%write(fd, tagLabels%egyTotal, energy%ETotal)
+    call taggedWriter%write(fd, tagLabels%fermiLvl, Ef)
+    call taggedWriter%write(fd, tagLabels%nElec, nEl)
 
     if (electronicSolver%providesEigenvals) then
       call taggedWriter%write(fd, tagLabels%freeEgy, energy%EMermin)
       ! extrapolated zero temperature energy
       call taggedWriter%write(fd, tagLabels%egy0Total, energy%Ezero)
+      call taggedWriter%write(fd, tagLabels%eigvals, eigen)
+      call taggedWriter%write(fd, tagLabels%eigFill, filling)
     end if
 
     if (tDefinedFreeE) then