+++ CHANGELOG.rst
@@ -31,7 +31,8 @@ Added
 
 - Non-SCC helical geometries supported
 
-- Generalized Born implicit solvation model
+- Generalized Born (GB) and Analytical Linearized Poisson-Boltzmann (ALPB)
+  implicit solvation model for SCC calculations
 
 - Non-polar solvent accessible surface area solvation model